Generalise CreateGeometry() and ProcessEvents() for any detector name.
The "-g" run option specifies the name of the Geant4 placement volume that
is used in CreateGeometry() to find the geometry parameters and boundaries.
The default name is "volArgonCubeDetector".

The "-d" run option specifies the name of the Geant4 sensitive detector
that contains the hits in ProcessEvents(). The default name is "ArgonCube".